Добавлены анимация и транзишны для верификации
